refactor perps components in TAP-1113
