+++ src/components/trade/left-content/perpetual/limit-order.tsx
@@ -1,71 +1,174 @@
-import { IUserStats } from '@/components/tapestry/models/drift.model'
-import { Input, Spinner, Switch } from '@/components/ui'
+import {
+  DirectionFilterType,
+  IUserStats,
+  LimitOrderParams,
+  OrderType,
+} from '@/components/tapestry/models/drift.model'
+import { useLeverageSize } from '@/components/trade/hooks/drift/use-leverage-size'
+import { useLiquidationPrice } from '@/components/trade/hooks/drift/use-liquidation-price'
+import { Input, Separator, Spinner, Switch } from '@/components/ui'
+import { formatUsdValue } from '@/utils/utils'
+import LeverageSelector from '@components/trade/left-content/perpetual/leverage-selector'
+import { Slippage } from '@components/trade/left-content/perpetual/slippage'
+import { ArrowRight } from 'lucide-react'
 import Image from 'next/image'
-import LeverageSelector from './leverage-selector'
-import { Slippage } from './slippage'
+import { useEffect, useState } from 'react'
 
-interface LimitOrderProps {
-  limitPrice: string
-  amount: string
+const SOL_IMG_URI =
+  'https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/So11111111111111111111111111111111111111112/logo.png'
+const USDC_IMG_URI =
+  'https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v/logo.png'
+
+interface Props {
+  symbol: string
+  priceLoading: boolean
   marketPrice: number
-  marketPriceLoading: boolean
-  leverageValue: number
-  selectedLeverageSizeUsd: string
-  selectedLeverageSizeToken: string
   userStats: IUserStats
-  swift: boolean
-  reduceOnly: boolean
-  isSizeByLeverage: boolean
-  slippageExpanded: boolean
-  slippageOption: string
-  handleLimitPriceChange: (e: React.ChangeEvent<HTMLInputElement>) => void
-  handleLeverageChange: (e: React.ChangeEvent<HTMLInputElement>) => void
-  handleAmountChange: (e: React.ChangeEvent<HTMLInputElement>) => void
-  handleDynamicSlippage: (e: React.ChangeEvent<HTMLInputElement>) => void
-  setSlippageExpanded: (value: boolean) => void
-  setSlippageOption: (value: string) => void
-  setAmount: (amount: string) => void
-  setLeverageValue: (leverage: number) => void
-  setSwift: (swift: boolean) => void
-  setReduceOnly: (reduceOnly: boolean) => void
-  setIsSizeByLeverage: (value: boolean) => void
+  direction: DirectionFilterType
+  setOrderParams: (params: LimitOrderParams) => void
 }
 
-const SOL_IMG_URI = "https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/So11111111111111111111111111111111111111112/logo.png"
-const USDC_IMG_URI = "https://raw.githubusercontent.com/solana-labs/token-list/main/assets/mainnet/EPjFWdd5AufqSSqeM2qN1xzybapC8G4wEGGkZwyTDt1v/logo.png"
-
 export default function LimitOrder({
-  limitPrice,
-  amount,
+  symbol,
+  direction,
+  priceLoading,
   marketPrice,
-  marketPriceLoading,
-  leverageValue,
-  selectedLeverageSizeUsd,
-  selectedLeverageSizeToken,
   userStats,
-  swift,
-  reduceOnly,
-  isSizeByLeverage,
-  slippageExpanded,
-  slippageOption,
-  handleLimitPriceChange,
-  handleAmountChange,
-  handleDynamicSlippage,
-  setAmount,
-  setSwift,
-  setReduceOnly,
-  setLeverageValue,
-  setIsSizeByLeverage,
-  setSlippageExpanded,
-  setSlippageOption
-}: LimitOrderProps) {
+  setOrderParams,
+}: Props) {
+  // States
+  const [amount, setAmount] = useState<string>('')
+  const [isSizeByLeverage, setIsSizeByLeverage] = useState<boolean>(false)
+  const [leverageValue, setLeverageValue] = useState<number>(1)
+  const [slippageOption, setSlippageOption] = useState<string>('0.1')
+  const [slippageExpanded, setSlippageExpanded] = useState<boolean>(false)
+  const [orderAmount, setOrderAmount] = useState<string>('')
+  const [limitPrice, setLimitPrice] = useState<string>('')
+  const [reduceOnly, setReduceOnly] = useState<boolean>(false)
+
+  const { selectedLeverageSizeUsd, selectedLeverageSizeToken } =
+    useLeverageSize({
+      userStats,
+      symbol,
+      leverageValue,
+      marketPrice: marketPrice,
+    })
+
+  const { liquidationPrice, loading: liqPriceLoading } = useLiquidationPrice({
+    symbol,
+    amount: orderAmount,
+    direction: direction === DirectionFilterType.LONG ? 'long' : 'short',
+  })
+  // handlers
+  const handleMarketOrderMarginChange = (
+    e: React.ChangeEvent<HTMLInputElement>
+  ) => {
+    const val = e.target.value
+
+    setIsSizeByLeverage(false)
+
+    if (
+      val === '' ||
+      val === '.' ||
+      /^[0]?\.[0-9]*$/.test(val) ||
+      /^[0-9]*\.?[0-9]*$/.test(val)
+    ) {
+      const maxAmount = userStats?.maxTradeSize || 0
+      const totalAmountInUSD = Number(val) * marketPrice
+      const newVal =
+        totalAmountInUSD > maxAmount
+          ? (maxAmount / marketPrice).toFixed(2)
+          : val
+
+      setAmount(newVal)
+
+      const cursorPosition = e.target.selectionStart
+      window.setTimeout(() => {
+        e.target.focus()
+        e.target.setSelectionRange(cursorPosition, cursorPosition)
+      }, 0)
+    }
+  }
+
+  const handleLimitPriceChange = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const val = e.target.value
+
+    if (
+      val === '' ||
+      val === '.' ||
+      /^[0]?\.[0-9]*$/.test(val) ||
+      /^[0-9]*\.?[0-9]*$/.test(val)
+    ) {
+      setLimitPrice(val)
+
+      const cursorPosition = e.target.selectionStart
+      window.setTimeout(() => {
+        e.target.focus()
+        e.target.setSelectionRange(cursorPosition, cursorPosition)
+      }, 0)
+    }
+  }
+
+  const handleDynamicSlippage = (e: React.ChangeEvent<HTMLInputElement>) => {
+    const val = e.target.value
+    const cursor = e.target.selectionStart || 0
+
+    if (
+      val === '' ||
+      val === '.' ||
+      /^[0]?\.[0-9]*$/.test(val) ||
+      /^[0-9]*\.?[0-9]*$/.test(val)
+    ) {
+      setSlippageOption(Number(val) < 100 ? val : '99')
+
+      const cursorPosition = e.target.selectionStart
+      window.setTimeout(() => {
+        e.target.focus()
+        e.target.setSelectionRange(cursorPosition, cursorPosition)
+      }, 0)
+    }
+  }
+
+  const formatLeverage = (leverage: number) => {
+    return leverage.toFixed(2) + 'x'
+  }
+
+  useEffect(() => {
+    if (!isSizeByLeverage) {
+      const newLeverageSize =
+        ((Number(amount) * marketPrice) / userStats.maxTradeSize) *
+        userStats.maxLeverage
+      setLeverageValue(newLeverageSize)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [amount, userStats, marketPrice])
+
+  useEffect(() => {
+    if (isSizeByLeverage) {
+      setOrderAmount(selectedLeverageSizeToken)
+    } else {
+      setOrderAmount(amount)
+    }
+    // eslint-disable-next-line react-hooks/exhaustive-deps
+  }, [amount, selectedLeverageSizeToken])
+
+  useEffect(() => {
+    setOrderParams({
+      orderType: OrderType.LIMIT,
+      amount: orderAmount,
+      slippage: slippageOption,
+      limitPrice,
+      reduceOnly,
+    } as LimitOrderParams)
+  }, [orderAmount, slippageOption, limitPrice, reduceOnly])
+
   return (
     <div className="space-y-4">
-      <div className='flex justify-between items-center'>
+      <div className="flex justify-between items-center">
         <span>Market Price</span>
         <div className="flex items-center gap-2">
-          <p className='text-primary'>{marketPrice.toFixed(4)} USD</p>
-          {marketPriceLoading && <Spinner size={12} />}
+          <p className="text-primary">{marketPrice.toFixed(4)} USD</p>
+          {priceLoading && <Spinner size={12} />}
         </div>
       </div>
 
@@ -91,7 +194,7 @@ export default function LimitOrder({
             placeholder="0.00"
             type="text"
             className="pr-12 text-primary text-xl"
-            onChange={(e) => handleAmountChange(e)}
+            onChange={(e) => handleMarketOrderMarginChange(e)}
             value={isSizeByLeverage ? selectedLeverageSizeToken : amount}
           />
           <Image
@@ -124,18 +227,18 @@ export default function LimitOrder({
         </div>
       </div>
 
-      {
-        userStats.maxLeverage ? (
-          <LeverageSelector
-            min={0}
-            max={Math.min(userStats.maxLeverage, 20)}
-            setAmount={setAmount}
-            leverageValue={leverageValue}
-            setLeverageValue={setLeverageValue}
-            setIsSizeByLeverage={setIsSizeByLeverage}
-          />
-        ) : (<></>)
-      }
+      {userStats.maxLeverage ? (
+        <LeverageSelector
+          min={0}
+          max={Math.min(userStats.maxLeverage, 20)}
+          setAmount={setAmount}
+          leverageValue={leverageValue}
+          setLeverageValue={setLeverageValue}
+          setIsSizeByLeverage={setIsSizeByLeverage}
+        />
+      ) : (
+        <></>
+      )}
 
       <Slippage
         handleDynamicSlippage={handleDynamicSlippage}
@@ -148,6 +251,31 @@ export default function LimitOrder({
       <div className="flex items-center space-x-2">
         <p>Reduce Only</p>
         <Switch checked={reduceOnly} onCheckedChange={setReduceOnly} />
+      </div>
+
+      <Separator />
+
+      <div className="flex justify-between items-center text-sm">
+        <p>Est.Liquidation Price</p>
+        <div className="flex items-center gap-2">
+          {liqPriceLoading ? (
+            <Spinner size={16} />
+          ) : liquidationPrice ? (
+            formatUsdValue(liquidationPrice)
+          ) : (
+            'None'
+          )}
+        </div>
+      </div>
+
+      <div className="flex justify-between items-center text-sm">
+        <p>Acct. Leverage</p>
+        <p className="flex items-center gap-2">
+          {formatLeverage(userStats.leverage)} <ArrowRight size={14} />
+          {formatLeverage(
+            userStats.leverage + (Number(amount) > 0 ? leverageValue : 0)
+          )}
+        </p>
       </div>
     </div>
   )
