+++ src/components/trade/left-content/perpetual/take-profit.tsx
