+++ src/components/trade/left-content/perpetual/updated-perpetual.tsx
@@ -31,7 +31,7 @@ import { Card, CardContent } from '@/components/ui/card'
 import { SOL_MINT } from '@/utils/constants'
 import { PositionDirection } from '@drift-labs/sdk-browser'
 import { useEffect, useMemo, useState } from 'react'
-import AddFundsModal from '../add-funds-modal'
+import AddFundsModal from './add-funds-modal'
 import LimitOrder from './limit-order'
 import MarketOrder from './market-order'
 import { PlacePerpsOrder } from './place-perps-order'
